refactor: notification register
