+++ src/endpoints/email/email.module.ts
@@ -4,10 +4,11 @@ import {
   MailModule,
   SubstrateModule,
 } from '../../common';
+import { NotificationModule } from '../notification/notification.module';
 import { EmailEndpointController } from './email.controller';
 
 @Module({
-  imports: [MailModule, SubstrateModule, EmailNotificationModule],
+  imports: [MailModule, SubstrateModule, EmailNotificationModule, NotificationModule],
   controllers: [EmailEndpointController],
 })
 export class EmailEndpointModule {}
